auth and post result filters
