--- Dayana/Server/Api/ResultFilters/Blog/PostResults/DeletePostResultFilter.cs
@@ -1,23 +1,0 @@
-﻿using Dayana.Shared.Basic.ConfigAndConstants.Constants.ConstMethods;
-using Dayana.Shared.Domains.Blog.BlogPosts;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-
-namespace Dayana.Server.Api.ResultFilters.Blog.PostResults;
-
-public class DeletePostResultFilter : ResultFilterAttribute
-{
-    public override async Task OnResultExecutionAsync(ResultExecutingContext context, ResultExecutionDelegate next)
-    {
-        var result = context.Result as ObjectResult;
-
-        if (result?.Value is Post value)
-            result.Value = new
-            {
-                Eid = value.Id.EncodeInt(),
-                value.PostTitle
-            };
-
-        await next();
-    }
-}
