--- Dayana/Server/Api/ResultFilters/Blog/PostResults/GetPostByFilterResultFilter.cs
@@ -1,33 +1,0 @@
-﻿using Dayana.Shared.Basic.ConfigAndConstants.Constants.ConstMethods;
-using Dayana.Shared.Infrastructure.Pagination;
-using Dayana.Shared.Persistence.Models.Blog.Base;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-
-namespace Dayana.Server.Api.ResultFilters.Blog.PostResults;
-
-public class GetPostByFilterResultFilter : ResultFilterAttribute
-{
-    public override async Task OnResultExecutionAsync(ResultExecutingContext context, ResultExecutionDelegate next)
-    {
-        var result = context.Result as ObjectResult;
-
-        if (result?.Value is PaginatedList<PostModel> value)
-            result.Value = new
-            {
-                value.Page,
-                value.PageSize,
-                value.TotalCount,
-                Data = value.Data.Select(x => new
-                {
-                    Eid = x.Id.EncodeInt(),
-                    x.PostTitle,
-                    x.Summery,
-                    x.PostBody,
-
-                })
-            };
-
-        await next();
-    }
-}
